+++ server/error-handler/index.js
@@ -3,8 +3,9 @@
  */
 
 const customError = require('./custom-error')
+const tools = require('./error-handler.tools')
 const middleware = require('./middleware')
 
 module.exports = Object.freeze(Object.assign({
   middleware
-}, customError))
+}, customError, tools))